Extract persistence code to new module, refine docstrings
